# docs/swagger.yaml
basePath: /
definitions:
  main.Subscription:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API для управления онлайн-подписками пользователей
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Возвращает все подписки с возможностью фильтрации по user_id и
        service_name
      parameters:
      - description: Фильтр по пользователю (UUID)
        in: query
        name: user_id
        type: string
      - description: Фильтр по названию сервиса
        in: query
        name: service_name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Subscription'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить список подписок
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Создаёт новую запись о подписке
      parameters:
      - description: Данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/main.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/main.Subscription'
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Создать подписку
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Удаляет подписку по ID
      parameters:
      - description: UUID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Подписка удалена
          schema:
            type: string
        "404":
          description: Подписка не найдена
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Удалить подписку
      tags:
      - subscriptions
    get:
      description: Возвращает подписку по её UUID
      parameters:
      - description: UUID подписки
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Subscription'
        "404":
          description: Подписка не найдена
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Получить подписку по ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Обновляет данные подписки по ID
      parameters:
      - description: UUID подписки
        in: path
        name: id
        required: true
        type: string
      - description: Новые данные подписки
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/main.Subscription'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.Subscription'
        "400":
          description: Неверный формат данных
          schema:
            type: string
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Обновить подписку
      tags:
      - subscriptions
  /subscriptions/summary:
    get:
      description: Возвращает сумму стоимости всех подписок с фильтрацией по user_id,
        service_name и периоду
      parameters:
      - description: UUID пользователя
        in: query
        name: user_id
        type: string
      - description: Название сервиса
        in: query
        name: service_name
        type: string
      - description: Начало периода (YYYY-MM)
        in: query
        name: start
        type: string
      - description: Конец периода (YYYY-MM)
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: total
          schema:
            additionalProperties:
              type: integer
            type: object
        "500":
          description: Ошибка сервера
          schema:
            type: string
      summary: Суммарная стоимость подписок
      tags:
      - subscriptions
swagger: "2.0"